+++ src/extractors/mod.rs
@@ -1,0 +1,2 @@
+pub mod links;
+pub mod wikitext;
