+++ src/extractors/wikitext/mod.rs
@@ -1,0 +1,3 @@
+mod template_transformers;
+pub mod tree;
+pub mod wikitext_parser;
