+++ src/lib.rs
@@ -1,7 +1,4 @@
-mod template_transformers;
-mod tree;
 mod work_queue;
-
-pub mod links;
-pub mod wikitext_parser;
 pub mod xml_parser;
+
+pub mod extractors;
